+++ K2D2Project/Controller/Lift/AutoLiftController.cs
@@ -1,3 +1,4 @@
+using K2D2.Controller.Lift.Pilots;
 using K2D2.KSPService;
 using KSP.Messages.PropertyWatchers;
 using KSP.Sim;
@@ -8,10 +9,6 @@ using Steamworks;
 using UnityEngine;
 
 namespace K2D2.Controller;
-
-
-
-
 
 public class AutoLiftController : ComplexController
 {
@@ -28,20 +25,35 @@ public class AutoLiftController : ComplexController
     public static AutoLiftController Instance { get; set; }
 
     AutoLiftSettings lift_settings = null;
-    LiftAscentPath ascentPath = null;
+    LiftAscentPath ascent_path = null;
 
     KSPVessel current_vessel;
 
-    float wanted_elevation;
+
+    Vector3d direction = Vector3d.zero;
+    bool show_profile = false;
+
+    // sub pilots
+    Ascent ascent = null;
+    Adjust adjust = null;
+    Coasting coasting = null;   
+    FinalCircularize final_circularize;
+
+    ExecuteController current_subpilot = null;
+
 
     public AutoLiftController()
     {
-        // logger.LogMessage("LandingController !");
         current_vessel = K2D2_Plugin.Instance.current_vessel;
 
         lift_settings = new AutoLiftSettings();
-        ascentPath = new LiftAscentPath(lift_settings);
+        ascent_path = new LiftAscentPath(lift_settings);
 
+        ascent = new Ascent(lift_settings, ascent_path);
+        adjust = new Adjust(lift_settings, ascent);
+        coasting = new Coasting(this, lift_settings);
+        final_circularize = new FinalCircularize(this, lift_settings);
+        
         Instance = this;
         debug_mode_only = false;
         name = "Lift";
@@ -69,74 +81,29 @@ public class AutoLiftController : ComplexController
                         // stop
                         if (current_vessel != null)
                             current_vessel.SetThrottle(0);
+
+                        current_subpilot = null;
                     }
                     break;
                 case LiftStatus.Ascent:
-                    {
-                        SASTool.setAutoPilot(AutopilotMode.StabilityAssist);
-                        last_ap_km = 0;
-                        ap_km = 0;
-                        delta_ap_per_second = 0;
-                        wanted_elevation = -90;
-                    }
+                    current_subpilot = ascent;
                     break;
                 case LiftStatus.Coasting:
-                    current_vessel.SetThrottle(0);
-                    turn_to = new TurnTo();
-                    turn_to.StartProGrade(SpeedDisplayMode.Surface);
-                    TimeWarpTools.SetRateIndex(0, false);
+                    current_subpilot = coasting;
                     break;
                 case LiftStatus.Adjust:
-                    current_vessel.SetThrottle(0);
-                    TimeWarpTools.SetRateIndex(0, false);
+                    current_subpilot = adjust;
                     break;
                 case LiftStatus.Circularize:
-                    {
-                        TimeWarpTools.SetRateIndex(0, false);
-                        if (K2D2OtherModsInterface.fpLoaded)
-                        {
-                            createCircleNode();
-                        }
-                        else
-                        {
-                            EndLiftPilot(true, "Please install FlightPlan for the final Step...");
-                        }
-                    }
+                    current_subpilot = final_circularize;
                     break;
             }
+
+            if (current_subpilot != null)
+                current_subpilot.Start();
         }
     }
 
-
-    void createCircleNode()
-    {
-        var current_time = GeneralTools.Game.UniverseModel.UniverseTime;
-
-        var current_vessel = K2D2_Plugin.Instance.current_vessel;
-        if (current_vessel == null)
-        {
-            EndLiftPilot(false, "error : no vessel");
-            return;
-        }
-
-        PatchedConicsOrbit orbit = current_vessel.VesselComponent.Orbit;
-        if (orbit == null)
-        {
-            EndLiftPilot(false, "error : no orbit");
-            return;
-        }
-
-        if (K2D2OtherModsInterface.instance.Circularize(current_time + orbit.TimeToAp, 0))
-        {
-            K2D2_Plugin.Instance.FlyNode();
-            EndLiftPilot(true, "Launched With FlightPlan !");
-        }
-        else
-        {
-            EndLiftPilot(false, "Error Creating Node");
-        }
-        return;
-    }
 
 
 
@@ -166,88 +133,23 @@ public class AutoLiftController : ComplexController
         status = LiftStatus.Ascent;
     }
 
-    Vector3d direction = Vector3d.zero;
-
-    public void applyDirection()
-    {
-        var autopilot = current_vessel.Autopilot;
-
-        if (autopilot == null)
-            return;
-
-        // force autopilot
-        autopilot.Enabled = true;
-
-        var telemetry = SASTool.getTelemetry();
-
-        var up = telemetry.HorizonUp;
-
-        direction = QuaternionD.Euler(-wanted_elevation, lift_settings.heading, 0) * Vector3d.forward;
-
-        Vector direction_vector = new Vector(up.coordinateSystem, direction);
-
-        autopilot.SAS.lockedMode = false;
-        autopilot.SAS.SetTargetOrientation(direction_vector, false);
-    }
-
-    float current_altitude_km = 0;
-    float ap_km = 0;
-    float last_ap_km = 0;
-    float delta_ap_per_second;
-    bool show_profile = false;
-
-    void computeValues()
-    {
-        if (current_vessel.VesselComponent == null)
-            return;
-
-        PatchedConicsOrbit orbit = current_vessel.VesselComponent.Orbit;
-        ap_km = (float)(orbit.Apoapsis - orbit.referenceBody.radius) / 1000;
-        current_altitude_km = (float)(current_vessel.GetSeaAltitude() / 1000);
-
-        if (last_ap_km == 0)
-        {
-            last_ap_km = ap_km;
-        }
-        else
-        {
-            float delta_ap = ap_km - last_ap_km;
-            last_ap_km = ap_km;
-            if (status == LiftStatus.Ascent)
-            {
-                // compute delta_ap_per_second only on ascent
-                float throttle = (float)current_vessel.GetThrottle();
-                if (throttle > 0.1f && Time.deltaTime != 0)
-                {
-                    float new_delta_ap_per_second = delta_ap / (Time.deltaTime * throttle);
-                    delta_ap_per_second = Mathf.Lerp(delta_ap_per_second, new_delta_ap_per_second, 0.1f);
-                }
-                else
-                    delta_ap_per_second = 0;
-            }
-        }
-
-        wanted_elevation = ascentPath.compute_elevation(current_altitude_km);
-    }
-
-    // used during coasting
-    double densityAtm = 0;
-    double duration_to_atm = 0;
-
-    TurnTo turn_to = null;
-
     bool result_ok = false;
     string end_status;
 
-    void EndLiftPilot(bool result_ok, string end_status)
+    public void EndLiftPilot(bool result_ok, string end_status)
     {
         this.result_ok = result_ok;
         this.end_status = end_status;
         status = LiftStatus.Off;
     }
 
-
     public WarpTo warp_to = new WarpTo();
+
+
+    public void NextMode()
+    {
+        status = status.Next();
+    }
 
     public override void Update()
     {
@@ -255,118 +157,36 @@ public class AutoLiftController : ComplexController
         if (current_vessel == null) return;
 
         if (isRunning || ui_visible)
-            computeValues();
+            ascent.computeValues(status == LiftStatus.Ascent);
 
         if (!isRunning)
             return;
 
-        switch (status)
+
+        if (current_subpilot != null)
         {
-            case LiftStatus.Coasting:
-                {
-                    CelestialBodyComponent mainBody = K2D2_Plugin.Instance.current_vessel.currentBody();
-                    var maxAtmosphereAltitude_km = (float)(mainBody.atmosphereDepth / 1000);
-                    if (lift_settings.destination_Ap_km < maxAtmosphereAltitude_km)
-                    {
-                        EndLiftPilot(false, "Warning Ap is under Atm. limit");
-                        return;
-                    }
-
-                    var altitude = (float)current_vessel.GetApproxAltitude() / 1000;
-                    densityAtm = mainBody.GetPressure(altitude * 1000);
-
-                    // compute time to reaching altitude.
-                    float V_Speed = (float)current_vessel.VesselVehicle.VerticalSpeed;
-                    var delta_alt = maxAtmosphereAltitude_km - altitude;
-
-                    if (delta_alt < 0)
-                    {
-                        // reached 
-                        status = LiftStatus.Adjust;
-                        return;
-                    }
-
-                    // warp until end
-                    duration_to_atm = delta_alt / V_Speed;
-                    var wanted_warp_index = WarpToSettings.compute_wanted_warp_index(duration_to_atm);
-                    TimeWarpTools.SetRateIndex(wanted_warp_index+1, false);
-                }
-                break;
-            case LiftStatus.Adjust:
-                {
-                    float remaining_Ap = lift_settings.destination_Ap_km - ap_km;
-                    if (remaining_Ap <= lift_settings.destination_Ap_km * 0.001f) // we stop at 0.1% of dest AP
-                    {
-                        status = LiftStatus.Circularize;
-                        return;
-                    }
-
-                    SASTool.setAutoPilot(AutopilotMode.Prograde);
-
-                    turn_to.Update();
-                    if (!turn_to.finished)
-                    {
-                        // adjust direction
-                        current_vessel.SetThrottle(0);
-                        return;
-                    }
-
-                    if (delta_ap_per_second == 0)
-                    {
-                        // set default ap per second to high value to get a smooth adjust
-                        delta_ap_per_second = 10;
-                    }
-
-                    wanted_throttle = remaining_Ap / delta_ap_per_second;
-                    wanted_throttle = wanted_throttle / 2;
-
-                    if (wanted_throttle > lift_settings.max_throttle)
-                        wanted_throttle = lift_settings.max_throttle;
-
-                    current_vessel.SetThrottle(wanted_throttle);
-                }
-                break;
-            case LiftStatus.Ascent:
-                {
-
-                    applyDirection();
-                    float remaining_Ap = lift_settings.destination_Ap_km - ap_km;
-                    if (remaining_Ap <= lift_settings.destination_Ap_km * 0.01f) // we stop at 0.1% of dest AP
-                    {
-                        status = LiftStatus.Coasting;
-                        return;
-                    }
-                    else
-                    {
-                        if (delta_ap_per_second <= 0)
-                        {
-                            wanted_throttle = lift_settings.max_throttle;
-                        }
-                        else
-                        {
-                            wanted_throttle = remaining_Ap / delta_ap_per_second;
-                            if (wanted_throttle > lift_settings.max_throttle)
-                                wanted_throttle = lift_settings.max_throttle;
-                        }
-
-                        current_vessel.SetThrottle(wanted_throttle);
-                    }
-                }
-                break;
+            current_subpilot.Update();
+            if (current_subpilot.finished)
+                NextMode();
         }
+
+
     }
 
-    float wanted_throttle = 0;
 
     public override void onGUI()
     {
-        UI_Tools.Title("Lift Pilot");
-
+        
         if (K2D2_Plugin.Instance.settings_visible)
         {
             K2D2Settings.onGUI();
+            lift_settings.onGUI();
+            K2D2Settings.CloseUI();
             return;
         }
+
+        //UI_Tools.Title("Lift Pilot");
+
 
         if (show_profile)
         {
@@ -384,7 +204,7 @@ public class AutoLiftController : ComplexController
             GUILayout.EndHorizontal();
 
             lift_settings.start_altitude_km = UI_Fields.IntFieldLine("lift.start_altitude_km", "90° Altitude", lift_settings.start_altitude_km, 0, Int32.MaxValue, "km");
-            ascentPath.drawProfile(current_altitude_km);
+            ascent_path.drawProfile(ascent.current_altitude_km);
 
             lift_settings.heading = UI_Tools.HeadingControl("lift.heading", lift_settings.heading);
 
@@ -415,42 +235,9 @@ public class AutoLiftController : ComplexController
         {
             UI_Tools.Warning($"Status : {status}");
 
-            switch (status)
-            {
-                case LiftStatus.Ascent:
-                    UI_Tools.Console($"Altitude = {current_altitude_km:n2} km");
-                    UI_Tools.Console($"Apoapsis Alt. = {ap_km:n2} km");
 
-                    UI_Tools.Console($"Inclination = {wanted_elevation:n2} °");
-
-                    UI_Tools.Console($"Last delta ap. = {delta_ap_per_second:n2} km/s");
-                    UI_Tools.Console($"wanted_throttle. = {wanted_throttle:n2}");
-                    break;
-                case LiftStatus.Adjust:
-                    UI_Tools.Console($"Altitude = {current_altitude_km:n2} km");
-                    UI_Tools.Console($"Apoapsis Alt. = {ap_km:n2} km");
-
-                    if (!turn_to.finished)
-                    {
-                        UI_Tools.Console(turn_to.status_line);
-                        return;
-                    }
-                    UI_Tools.Console($"wanted_throttle. = {wanted_throttle:n2}");
-                    break;
-                case LiftStatus.Coasting:
-                    UI_Tools.Console($"Altitude = {current_altitude_km:n2} km");
-                    UI_Tools.Console($"Atm Density = {densityAtm:n2}");
-                    UI_Tools.Console($"End warp : {StrTool.DurationToString(duration_to_atm)} x{TimeWarpTools.CurrentRate}");
-                    break;
-                case LiftStatus.Circularize:
-                    UI_Tools.Console("--------------");
-                    if (UI_Tools.Button("Final !"))
-                    {
-                        createCircleNode();
-                    }
-
-                    break;
-            }
+            if (current_subpilot != null)
+                current_subpilot.onGUI();
         }
         else
         {
@@ -458,11 +245,11 @@ public class AutoLiftController : ComplexController
             {
                 if (result_ok)
                 {
-                    UI_Tools.OK(end_status);
+                    UI_Tools.OK("Final status : " + end_status);
                 }
                 else
                 {
-                    UI_Tools.Warning(end_status);
+                    UI_Tools.Warning("Final status : " + end_status);
                 }
             }
         }